+++ src/main/java/be/baur/sds/content/BooleanType.java
@@ -1,13 +1,13 @@
 package be.baur.sds.content;
 
-import be.baur.sds.SimpleType;
+import be.baur.sds.NodeType;
 import be.baur.sds.common.Content;
 
 /**
  * A <code>BooleanType</code> is a simple type representing an SDA node with
  * boolean content, e.g. "true" or "false". No other values are allowed.
  */
-public final class BooleanType extends SimpleType {
+public final class BooleanType extends NodeType {
 
 	public static final String TRUE = "true";
 	public static final String FALSE = "false";